Fixed key erro, refactored contants out of Model
