+++ client/components/modal/ModalModel.jsx
@@ -3,51 +3,11 @@ import PropTypes from 'prop-types';
 import Stars from '../util/Stars';
 import ProductInfo from './ProductInfo';
 import ModalHeader from './ModalHeader';
-import UserProductRating from './UserProductRating';
-import ModalSubmit from './ModalSubmit';
-import UserText from './UserText';
-import UserButton from './UserButton';
 import ModalContainer from './ModalContainer';
-
-const TITLES = [
-  'Product Rating',
-  'Review title',
-  'Review',
-  'Would you recommend this product to a friend?',
-  'Nickname',
-  'Location',
-  'Email',
-  'Fit',
-  'Did you read product reviews online before first purchasing this item?',
-  'Where did you purchase this item?',
-  'What feedback do you have for the people who designed and manufactured this product?',
-  'Submit',
-];
-
-const SAMPLES = [
-  'Makes hiking even easier',
-  'jackie27',
-  'Seattle, WA',
-  'youremail@example.com',
-];
+import MODALS from './ModalConstants';
 
 const REQ = [
   true, true, true, false, true, false, true, false, false, false, true,
-];
-
-const MODALS = [
-  (<UserProductRating />),
-  (<UserText title={TITLES[1]} required text placeholder={SAMPLES[0]} />),
-  (<UserText title={TITLES[2]} required text={false} placeholder="" />),
-  (<UserButton title={TITLES[3]} />),
-  (<UserText title={TITLES[4]} required text placeholder={SAMPLES[1]} />),
-  (<UserText title={TITLES[5]} required text placeholder={SAMPLES[2]} />),
-  (<UserText title={TITLES[6]} required text placeholder={SAMPLES[3]} />),
-  (<UserButton title={TITLES[7]} />),
-  (<UserButton title={TITLES[8]} />),
-  (<UserButton title={TITLES[9]} />),
-  (<UserText title={TITLES[10]} required text={false} placeholder="" />),
-  (<ModalSubmit />),
 ];
 
 class ModalModel extends Component {
@@ -113,7 +73,7 @@ class ModalModel extends Component {
               {MODALS.map((modal, index) => (
                 <ModalContainer
                   active={active === index}
-                  key={TITLES[index]}
+                  key={modal.toString() + index}
                   required={REQ[index]}
                   submit={submit}
                   hasValue={values[index] !== ''}
